multi: cleaned up files

+++ src/driver.cpp
@@ -24,7 +24,7 @@ driver::~driver()
 }
 
 // METHODS
-void driver::initialize(unsigned int gpio_pin_a, unsigned int gpio_pin_b, unsigned int ppr, double spin_ratio)
+void driver::initialize(uint32_t gpio_pin_a, uint32_t gpio_pin_b, uint32_t ppr, double spin_ratio)
 {
     // Store the cpr and spin_ratios.
     driver::m_cpr = ppr * 2UL;  // Two counts for each pulse (rising/falling edge).
@@ -53,23 +53,23 @@ double driver::get_position()
 void driver::tick_a(bool level)
 {
     // Calculate the new state.
-    unsigned int new_state = (driver::m_prior_state & 1) | (static_cast<unsigned int>(level) << 1);
+    uint32_t new_state = (driver::m_prior_state & 1) | (static_cast<uint32_t>(level) << 1);
     // Update the encoder's state with the new state.
     driver::update_state(new_state);
 }
 void driver::tick_b(bool level)
 {
     // Calculate the new state.
-    unsigned int new_state = (driver::m_prior_state & 2) | static_cast<unsigned int>(level);
+    uint32_t new_state = (driver::m_prior_state & 2) | static_cast<uint32_t>(level);
     // Update the encoder's state with the new state.
     driver::update_state(new_state);
 }
 void driver::initialize_state(bool level_a, bool level_b)
 {
     // Initialize the prior state with the current state.
-    driver::m_prior_state = (static_cast<unsigned int>(level_a) << 1) | static_cast<unsigned int>(level_b);
+    driver::m_prior_state = (static_cast<uint32_t>(level_a) << 1) | static_cast<uint32_t>(level_b);
 }
-void driver::update_state(unsigned int new_state)
+void driver::update_state(uint32_t new_state)
 {
     // Get the transition value.
     int transition = driver::m_transition_matrix.at(driver::m_prior_state).at(new_state);
@@ -89,7 +89,7 @@ void driver::update_state(unsigned int new_state)
     driver::m_prior_state = new_state;
 }
 
-long long int driver::p_pulses_missed()
+uint64_t driver::p_pulses_missed()
 {
     return driver::m_pulses_missed;
 }
